+++ src/panels/lovelace/editor/entity-row-editor/hui-row-element-editor.ts
@@ -2,12 +2,12 @@ import { customElement } from "lit/decorators";
 import { getRowElementClass } from "../../create-element/create-row-element";
 import { LovelaceRowConfig } from "../../entity-rows/types";
 import type { LovelaceRowEditor } from "../../types";
-import { HuiElementEditor } from "../hui-element-editor";
+import { HuiElementTypeEditor } from "../hui-element-type-editor";
 
 const GENERIC_ROW_TYPE = "generic-row";
 
 @customElement("hui-row-element-editor")
-export class HuiRowElementEditor extends HuiElementEditor<LovelaceRowConfig> {
+export class HuiRowElementEditor extends HuiElementTypeEditor<LovelaceRowConfig> {
   protected get configElementType(): string | undefined {
     if (!this.value?.type && "entity" in this.value!) {
       return GENERIC_ROW_TYPE;
