+++ src/panels/lovelace/editor/types.ts
@@ -9,6 +9,7 @@ import { LovelaceHeaderFooterConfig } from "../header-footer/types";
 import { LovelaceCardFeatureConfig } from "../card-features/types";
 import { LovelaceElementConfig } from "../elements/types";
 import { LovelaceBadgeConfig } from "../../../data/lovelace/config/badge";
+import { HeadingCardEntityConfig } from "../cards/types";
 
 export interface YamlChangedEvent extends Event {
   detail: {
@@ -95,7 +96,8 @@ export interface SubElementEditorConfig {
     | LovelaceRowConfig
     | LovelaceHeaderFooterConfig
     | LovelaceCardFeatureConfig
-    | LovelaceElementConfig;
+    | LovelaceElementConfig
+    | HeadingCardEntityConfig;
   type: "header" | "footer" | "row" | "feature" | "element" | "heading-entity";
 }
